+++ src/@core/application/useCases/product/ListProductsUseCase.ts
@@ -1,6 +1,5 @@
 import { IListProductsUseCase, Product, ProductCategory } from '@core/domain';
 import { Injectable } from '@nestjs/common';
-import { Decimal } from '@prisma/client/runtime/library';
 import { PrismaService } from '@prismaOrm/prisma.service';
 
 @Injectable()
@@ -13,7 +12,7 @@ export class ListProductsUseCase implements IListProductsUseCase {
     return products.map((p) => ({
       ...p,
       category: p.category as ProductCategory,
-      price: (p.price as Decimal).toNumber(),
+      price: p.price.toNumber(),
     }));
   }
 }
